Refactor Serena tool creation
